requirements.txt
opencv-python-headless==4.9.0.80
tqdm==4.67.1
imageio==2.36.0
einops==0.8.0
scipy==1.14.1
matplotlib==3.9.2
wandb==0.18.7
timm==1.0.7
ninja==1.11.1
numpy==1.26.3
huggingface-hub==0.23.4
kornia==0.7.4
torch-scatter==2.1.2